Call setVisible(true) after components are added
I was having a problem where no elements in the frame were visible.
I've found out that if I resize the window they show up and after that it works.

The correct way to make this work all the time is to call the setVisible(true) method only after the elements are added to the frame.

Some more on that: http://stackoverflow.com/a/12713622/679227.

See: https://github.com/zonama-sekot/zonama-fighter/pull/7